ConversationDialogue UI Art Prefab
Developed prefabs for the conversation dialogue background, next and back buttons, and the character name plate with temporary text.
Designed a subgraph for light sweeps for UI assets and created a light sweep material to add a looping shine to the outer trim of the conversation dialogue background. This will serve as a template for looping shines on other UI/menu prefabs.
Constructed an animator component, animation clips, and triggers for button idle, mouse hover, and click states with animated transitions (Idle>HoverIn>HoverHold>Click>HoverOut).